Refactor timer
