+++ exercises/fulfill_a_promise/exercise.js
@@ -5,8 +5,7 @@ var filecheck = require('workshopper-exercise/filecheck');
 var execute = require('workshopper-exercise/execute');
 var comparestdout = require('workshopper-exercise/comparestdout');
 var wrappedexec = require('@timothygu/workshopper-wrappedexec');
-var starttime = require('../../lib/starttime');
-var endtime = require('../../lib/endtime');
+var timer = require('../../lib/timer');
 
 var THRESHOLD = 100;
 
@@ -16,11 +15,8 @@ exercise = filecheck(exercise);
 // execute the solution and submission in parallel with spawn()
 exercise = execute(exercise);
 
-// start timer
-exercise = starttime(exercise);
-
-// end timer
-exercise = endtime(exercise, THRESHOLD);
+// timer
+exercise = timer(exercise, THRESHOLD);
 
 // compare stdout of solution and submission
 exercise = comparestdout(exercise);
