+++ exercises/warm_up/exercise.js
@@ -4,8 +4,7 @@ var exercise = require('workshopper-exercise')();
 var filecheck = require('workshopper-exercise/filecheck');
 var execute = require('workshopper-exercise/execute');
 var comparestdout = require('workshopper-exercise/comparestdout');
-var starttime = require('../../lib/starttime');
-var endtime = require('../../lib/endtime');
+var timer = require('../../lib/timer.js');
 
 var THRESHOLD = 100;
 
@@ -15,11 +14,8 @@ exercise = filecheck(exercise);
 // execute the solution and submission in parallel with spawn()
 exercise = execute(exercise);
 
-// start timer
-exercise = starttime(exercise);
-
-// end timer
-exercise = endtime(exercise, THRESHOLD);
+// timer
+exercise = timer(exercise, THRESHOLD);
 
 // compare stdout of solution and submission
 exercise = comparestdout(exercise);
