+++ lib/timer.js
@@ -3,6 +3,19 @@
 var through2 = require('through2');
 var bl = require('bl');
 
+function setup(mode, callback) {
+  this.start = {
+    submission: new Date(),
+    solution: new Date(),
+  };
+  this.duration = {
+    submission: 0,
+    solution: 0,
+  };
+
+  process.nextTick(callback);
+}
+
 function verifyProcessor(threshold, callback) {
   function verify(cur, alt) {
     var real = this[cur + 'Stdout'];
@@ -37,10 +50,11 @@ function verifyProcessor(threshold, callback) {
   });
 }
 
-function endtime(exercise, threshold, minimum) {
+function timer(exercise, threshold) {
+  exercise.addSetup(setup);
   exercise.addVerifyProcessor(verifyProcessor.bind(exercise, threshold));
 
   return exercise;
 }
 
-module.exports = endtime;
+module.exports = timer;
